# cut heap allocation by 23%